supabase_py_1075/pyproject.toml
[project]
name = "supabase-py-1075"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "python-dotenv>=1.0.1",
    "supabase>=2.14.0",
]
